Strip tests to the barebones
