--- src/test/java/com/gregtechceu/gtceu/gametest/Batches.java
@@ -1,5 +1,0 @@
-package com.gregtechceu.gtceu.gametest;
-
-public final class Batches {
-
-}
